+++ src/App.js
@@ -1,15 +1,16 @@
 import React, {useRef, useState } from 'react';
-import './App.css';
+import './styles/App.css';
+import "./styles/styles.css";
 
 import { initializeApp } from "firebase/app"
-import { getAuth, signInWithEmailAndPassword, GoogleAuthProvider, signInWithPopup, createUserWithEmailAndPassword, signOut } from 'firebase/auth';
+import { getAuth, signOut} from 'firebase/auth';
 import { setDoc, getDocs, getFirestore, collection, deleteDoc, orderBy, limit, doc, serverTimestamp, query, where} from "firebase/firestore";
-
-import "./styles.css";
-import LoginForm from "./LoginForm";
 
 import { useAuthState } from 'react-firebase-hooks/auth';
 import { useCollectionData } from 'react-firebase-hooks/firestore';
+
+
+import SignIn from './components/SignIn';
 
 
 const firebaseApp = initializeApp({
@@ -38,97 +39,11 @@ function App() {
       </header>
 
       <section>
-        {user ? <ChatRoom /> : <SignIn />}
+        {user ? <ChatRoom /> : <SignIn auth = {auth}/>}
       </section>
 
     </div>
   );
-}
-
-
-
-function SignIn() {
-
-  const [isShowLogin, setIsShowLogin] = useState(true);
-  const [isShowSignup, setIsShowSignup] = useState(true);
-
-
-
-  const signInWithGoogle = async () => {
-    const provider = new GoogleAuthProvider();
-    await signInWithPopup(auth, provider).then((result) => {
-    }).catch((error) => {
-      var errorCode = error.code;
-      var errorMessage = error.message;
-      document.getElementById("errorMessage").innerHTML = "<h2>"+errorCode + " " + errorMessage + "</h2>";
-    }
-    );
-  }
-
-
-  const logInPop = () => {
-    if(!isShowSignup){
-      setIsShowSignup((isShowSignup) => !isShowSignup);
-    }
-      setIsShowLogin((isShowLogin) => !isShowLogin);
-  }
-
-  const SignUpPop = () => {
-    if(!isShowLogin){
-      setIsShowLogin((isShowLogin) => !isShowLogin);
-    }
-    setIsShowSignup((isShowSignup) => !isShowSignup);   
-  }
-
-  const signInWithEmail = () => {
-    const email = document.getElementById("usernamesignin").value;
-    const password = document.getElementById("passwordsignin").value;
-    signInWithEmailAndPassword(auth, email, password).then((userCredential) => {
-      // Signed in
-      var user = userCredential.user;
-    })
-    .catch((error) => {
-      var errorCode = error.code;
-      var errorMessage = error.message;
-      document.getElementById("errorMessage").innerHTML = "<h2>"+errorCode + " " + errorMessage + "</h2>";
-    });
-
-    setIsShowLogin((isShowLogin) => !isShowLogin);
-  }
-
-  const signUpWithEmail = () => {
-    const email = document.getElementById("usernamesignup").value;
-    const password = document.getElementById("passwordsignup").value;
-    createUserWithEmailAndPassword(auth, email, password).then((userCredential) => {
-      // Signed in
-      var user = userCredential.user;
-    }
-    ).catch((error) => {
-      var errorCode = error.code;
-      var errorMessage = error.message;
-      document.getElementById("errorMessage").innerHTML = "<h2>"+errorCode + " " + errorMessage + "</h2>";
-    });
-    setIsShowSignup((isShowSignup) => !isShowSignup);
-  }
-
-  return (
-    <>
-      <h1 style={{ color: 'orange' }}>Chat</h1>
-      <div id='errorMessage'></div>
-      <button className="sign-in" onClick={logInPop}>Sign in with Email</button>
-      <br></br>
-      <button className="sign-in" onClick={signInWithGoogle}>Sign in with Google</button>
-      <br></br>
-      <a style={{ color: 'white' }} onClick={SignUpPop} >Registarse</a>
-      <div className="LoginPopUp">
-         <LoginForm isShowLogin={isShowLogin} submit={signInWithEmail} searchId="signin" topText="Sign In" />
-         <LoginForm isShowLogin={isShowSignup} submit={signUpWithEmail} searchId="signup" topText="Sign Up" />
-       </div>
-  
-        
-      </>
-  )
-
 }
 
 function SignOut() {
